+++ src/SimShimPlotCtx.js
@@ -1,6 +1,13 @@
 import THREE from 'three'
 
 
+/**
+ * This is a container class that holds everything needed for rendering (the
+ * scene, camera, renderer, list of SimShimObjs, lights, ...) and also
+ * some meta info about the scene like the bounding box around all geometries
+ * it is holding (for automatic camera targeting).
+ *
+ */
 export default class SimShimPlotCtx {
   constructor (renderer, scene, camera, controls, light) {
     this.renderer = renderer;
@@ -8,11 +15,11 @@ export default class SimShimPlotCtx {
     this.camera   = camera;
     this.controls = controls;
     this.light    = light;
-    this.plots    = [];
+    this.objects  = [];
   }
 
   updateMetrics () {
-    // init (also defaults for when no plots exist)
+    // init (also defaults for when no objects exist)
     var res = {
       'maxX': 0, 'maxY': 0, 'maxZ': 0,
       'minX': 0, 'minY': 0, 'minZ': 0,
@@ -37,13 +44,12 @@ export default class SimShimPlotCtx {
       }
     }
 
-    // if plots exist, update these values
-    if (this.plots.length) {
+    // if objects exist, update these values
+    if (this.objects.length) {
 
       // iterate
-      this.plots.forEach((plt) => {
-        if (plt.threeObj) setMaxMin( plt.threeObj.geometry.vertices );
-        else console.warn('SimShimPlot found with no THREEjs object');
+      this.objects.forEach((ssObj) => {
+        if (ssObj.threeObj) setMaxMin( ssObj.threeObj.geometry.vertices )
       });
 
       // compute extra metrics