+++ CHANGELOG.md
@@ -1,8 +1,19 @@
+**0.2.0**
+
+* Add basic documentation (ESDoc.org)
+* Add lots of helpful errors
+* Rename `SimShimPlot` to `SimShimObj`
+* Rename `getPlot` to `getById`
+* Internal API refactor, breaking changes for advanced usage (simple usages,
+  like those found in the `examples/` folder, are unaffected) -- API for
+  SimShimPlot has changed, see docs
+
+
 **0.1.0**
 
-* npm publish, moved all dependencies to npm
+* Getting ready for npm publish, moved all dependencies to npm
 
 
-**Before npm publishing...**
+**0.0.0**
 
-* instance method `kill` has been rewritten and renamed to `removeAllObjects`
+* Instance method `kill` has been rewritten and renamed to `removeAllObjects`